[Add] hm00 flatbuffer specification for N-D histograms
Add features from hs00/hs01 into a structure akin to ADAr.
The histogram itself is moved into a sub-table so that it can be used
for both a required data histogram and optional errors histogram.

Optional bin-boundary specifications are added. The attributes from ADAr
is ported over.

By moving repeated object-types into tables, dataclasses can be used in
Python to more-easily serialise and deserialise the flatbuffer.

(A python implentation of the serialiser and deserialiser for hm01 has
been written, with inspiration taken from ADAr)